Reorganise the current code
