+++ src/pages/Login/components/LoginForm/LoginForm.tsx
@@ -1,0 +1,115 @@
+import React from "react";
+import Avatar from "@mui/material/Avatar";
+import Button from "@mui/material/Button";
+import TextField from "@mui/material/TextField";
+import FormControlLabel from "@mui/material/FormControlLabel";
+import Checkbox from "@mui/material/Checkbox";
+import Link from "@mui/material/Link";
+import Paper from "@mui/material/Paper";
+import Box from "@mui/material/Box";
+import Grid from "@mui/material/Grid";
+import LockOutlinedIcon from "@mui/icons-material/LockOutlined";
+import Typography from "@mui/material/Typography";
+import GoogleIcon from "@mui/icons-material/Google";
+import LoginIcon from "@mui/icons-material/Login";
+import LogInLogic from "../../services/email-login";
+import { Copyright } from "../../../../components/Copyright";
+import GoogleLogin from "../../services/google-login";
+
+const LoginForm = () => {
+  return (
+    <Grid
+      item
+      xs={12}
+      sm={12}
+      md={5}
+      component={Paper}
+      elevation={6}
+      square
+      sx={{ width: "100%" }}
+    >
+      <Box
+        sx={{
+          my: 8,
+          mx: 4,
+          display: "flex",
+          flexDirection: "column",
+          alignItems: "center",
+        }}
+      >
+        <Avatar sx={{ m: 1, bgcolor: "secondary.main" }}>
+          <LockOutlinedIcon />
+        </Avatar>
+        <Typography component="h1" variant="h5">
+          Sign In
+        </Typography>
+        <Box component="form" noValidate sx={{ mt: 1 }}>
+          <Grid container spacing={2}>
+            <Grid item xs={12}>
+              <TextField
+                required
+                fullWidth
+                id="email"
+                label="Email Address"
+                name="email"
+                autoComplete="email"
+              />
+            </Grid>
+            <Grid item xs={12}>
+              <TextField
+                required
+                fullWidth
+                name="password"
+                label="Password"
+                type="password"
+                id="password"
+                autoComplete="new-password"
+              />
+            </Grid>
+          </Grid>
+
+          <FormControlLabel
+            control={<Checkbox value="remember" color="primary" />}
+            label="Remember me"
+          />
+          <Button
+            type="submit"
+            fullWidth
+            variant="contained"
+            startIcon={<LoginIcon />}
+            sx={{ mt: 3, mb: 1 }}
+            onClick={LogInLogic}
+          >
+            Sign In
+          </Button>
+
+          <Button
+            fullWidth
+            variant="contained"
+            startIcon={<GoogleIcon />}
+            style={{ borderColor: "#DB4437", backgroundColor: "#DB4437" }}
+            sx={{ mb: 2 }}
+            onClick={GoogleLogin}
+          >
+            Sign In with Google
+          </Button>
+          <Grid container>
+            <Grid item xs>
+              <Link href="#" variant="body2">
+                Forgot password?
+              </Link>
+            </Grid>
+            <Grid item>
+              <Link href="/signup" variant="body2">
+                {"Don't have an account? Sign Up"}
+              </Link>
+            </Grid>
+          </Grid>
+          <Copyright sx={{ mt: 5 }} />
+        </Box>
+      </Box>
+    </Grid>
+  );
+};
+
+export default LoginForm;
