+++ src/pages/Login/services/google-login.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+import { GoogleAuthProvider, signInWithPopup } from "firebase/auth";
+import { useState } from "react";
+import { auth } from "../../../firebase/firebase";
+
+const googleLogin = () => {
+  const provider = new GoogleAuthProvider();
+
+  return signInWithPopup(auth, provider).then(() => {
+    window.location.href = "/";
+  });
+};
+
+export default googleLogin;
