+++ src/pages/Signup/components/SignUpForm/SignUpForm.tsx
@@ -1,0 +1,130 @@
+import React from "react";
+import Avatar from "@mui/material/Avatar";
+import Button from "@mui/material/Button";
+import TextField from "@mui/material/TextField";
+import FormControlLabel from "@mui/material/FormControlLabel";
+import Checkbox from "@mui/material/Checkbox";
+import Link from "@mui/material/Link";
+import Paper from "@mui/material/Paper";
+import Box from "@mui/material/Box";
+import Grid from "@mui/material/Grid";
+import LockOutlinedIcon from "@mui/icons-material/LockOutlined";
+import Typography from "@mui/material/Typography";
+import GoogleIcon from "@mui/icons-material/Google";
+import LockOpenIcon from "@mui/icons-material/LockOpen";
+import { Copyright } from "../../../../components/Copyright";
+import GoogleSignUp from "../../services/GoogleSignup";
+import SignUpLogic from "../../services/SignUpLogic";
+
+const SignUpForm = () => {
+  return (
+    <Grid
+      item
+      xs={12}
+      sm={12}
+      md={5}
+      component={Paper}
+      elevation={6}
+      square
+      sx={{ width: "100%" }}
+    >
+      <Box
+        sx={{
+          my: 8,
+          mx: 4,
+          display: "flex",
+          flexDirection: "column",
+          alignItems: "center",
+        }}
+      >
+        <Avatar sx={{ m: 1, bgcolor: "secondary.main" }}>
+          <LockOutlinedIcon />
+        </Avatar>
+        <Typography component="h1" variant="h5">
+          Sign Up
+        </Typography>
+        <Box component="form" noValidate sx={{ mt: 1 }}>
+          <Grid container spacing={2}>
+            <Grid item xs={12} sm={6}>
+              <TextField
+                autoComplete="given-name"
+                name="firstName"
+                required
+                fullWidth
+                id="firstName"
+                label="First Name"
+                autoFocus
+              />
+            </Grid>
+            <Grid item xs={12} sm={6}>
+              <TextField
+                required
+                fullWidth
+                id="lastName"
+                label="Last Name"
+                name="lastName"
+                autoComplete="family-name"
+              />
+            </Grid>
+            <Grid item xs={12}>
+              <TextField
+                required
+                fullWidth
+                id="email"
+                label="Email Address"
+                name="email"
+                autoComplete="email"
+              />
+            </Grid>
+            <Grid item xs={12}>
+              <TextField
+                required
+                fullWidth
+                name="password"
+                label="Password"
+                type="password"
+                id="password"
+                autoComplete="new-password"
+              />
+            </Grid>
+          </Grid>
+          <FormControlLabel
+            control={<Checkbox value="remember" color="primary" />}
+            label="Remember me"
+          />
+          <Button
+            type="submit"
+            fullWidth
+            variant="contained"
+            startIcon={<LockOpenIcon />}
+            sx={{ mt: 3, mb: 1 }}
+            onClick={SignUpLogic}
+          >
+            Sign Up with Form
+          </Button>
+
+          <Button
+            fullWidth
+            variant="contained"
+            startIcon={<GoogleIcon />}
+            style={{ borderColor: "#DB4437", backgroundColor: "#DB4437" }}
+            sx={{ mb: 2 }}
+            onClick={GoogleSignUp}
+          >
+            Sign Up with Google
+          </Button>
+          <Grid container>
+            <Grid item>
+              <Link href="/login" variant="body2">
+                {"Already have an account? Sign in"}
+              </Link>
+            </Grid>
+          </Grid>
+          <Copyright sx={{ mt: 5 }} />
+        </Box>
+      </Box>
+    </Grid>
+  );
+};
+
+export default SignUpForm;
